.specify/physical-ai-rag-chatbot/backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
pydantic==2.9.2
python-dotenv==1.0.1
qdrant-client==1.12.0
sqlalchemy==2.0.36
asyncpg==0.30.0
psycopg[binary]==3.2.1
pyjwt==2.9.0
cryptography==43.0.1
langchain==0.3.0
langchain-community==0.3.0
langchain-google-genai==1.0.8
google-generativeai==0.8.4